Switch to external parameter settings
This switches to a model where parameters are set from outside of
instrument implementations, and instruments are simply told that the
values of parameters have changed. The values of parameters are changed
in a manner that is timing correct (that is, an instrument won't see
that the value of a parameter has changed until it is supposed to
have happened at the correct time within a processing period).

This eliminates all parameter set message types, allows for hosts
to handle the saving/loading of state, and leaves the door open for
instruments to be able to read suitable parameters at audio rate
if desired.